requirements.txt
xformers==0.0.19
triton==2.0.0
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.0+cu118
torchvision==0.15.1+cu118
torchaudio==2.0.1+cu118
torchtext==0.15.1
torchdata==0.6.0
PyYAML==6.0